Enemy Spawn, Main Menu and more
I'm starting to lose track of what I'm working on, I've gotten a bit sloppy of checking what I still need to work on, what I have worked on and what I'm yet to touch, I had to take a break from the project for a week, and that's screwed me over. This push is mainly for version checking, and allowing me to see what I've changed on GitHub, as it's easier than the desktop app.